CSV Reading : One line at a time using custom  CSV_Reader library utilizing iterators
Previously, the whole CSV file's rows (each of which was a list) was
saved into a list called ${InvalidLoginScenarios}. This was inefficient,
as the file size grows.

Now we use a generator called return_value() which returns the rows one
at a time when next() is called. When all rows are read, the generator
throws StopIteration.

CSV_Reader is a custom library object, holding state information (i.e.
_csvfile and _reader). Note that _reader is the one keeping track of the
next row to be read.

When read_row_from_csv_file() is called, it gets an iterator, which
holds the state information (i.e. the next row to be read) via _reader.
if next() call throws StopIteration error, that means we finished
reading the file. Otherwise, we return the row read.