CSV Reading : One Line At A Time : Improved
note that self._reader is an iterable. If we call iter() on it, we will
get an iterator. So, no need to create an additional generator/iterator
which will iterate of self._iter via a for loop.

The uncessary for loop would extend to:

iterator = iter(reader)
while True:
	try:
		row = next(iterator)
        except StopIteration:
		break
	else
		yield row
pass 	# when execution reaches here, the generator will throw
StopIteration error